+++ include/H2DE/engine/H2DE_renderer.h
@@ -66,10 +66,10 @@ private:
 
     static constexpr float getOpacityBlend(uint8_t opacity) { return (static_cast<float>(opacity) / static_cast<float>(H2DE_UINT8_MAX)); }
 
-    H2DE_PixelPos levelToPixelPos(const H2DE_LevelRect& world_rect, bool absolute) const;
-    H2DE_PixelPos levelToPixelPos(const H2DE_Translate& local_translate, bool absolute) const;
-    H2DE_PixelSize levelToPixelSize(const H2DE_Scale& world_scale, bool absolute) const;
-    H2DE_PixelRect levelToPixelRect(const H2DE_LevelRect& world_rect, bool absolute) const;
+    H2DE_SubPixelPos levelToPixelPos(const H2DE_LevelRect& world_rect, bool absolute) const;
+    H2DE_SubPixelPos levelToPixelPos(const H2DE_Translate& local_translate, bool absolute) const;
+    H2DE_SubPixelSize levelToPixelSize(const H2DE_Scale& world_scale, bool absolute) const;
+    H2DE_SubPixelRect levelToPixelRect(const H2DE_LevelRect& world_rect, bool absolute) const;
 
     H2DE_Translate pixelToLevel(const H2DE_PixelPos& pos, bool absolute) const;